move address-info metric to netdev_common
